Collection owner

+++ src/main/java/ua/com/javarush/parse/m5/passwordmanager/service/CollectionService.java
@@ -8,6 +8,7 @@ import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import ua.com.javarush.parse.m5.passwordmanager.entity.Collection;
+import ua.com.javarush.parse.m5.passwordmanager.entity.User;
 import ua.com.javarush.parse.m5.passwordmanager.repository.CollectionRepository;
 
 @Service
@@ -15,40 +16,63 @@ import ua.com.javarush.parse.m5.passwordmanager.repository.CollectionRepository;
 public class CollectionService {
 
   private final CollectionRepository collectionRepository;
+  private final UserService userService;
 
-  @Cacheable(value = "collections", key = "'all'")
+  @Cacheable(
+      value = "collections",
+      key =
+          "'user:' + T(org.springframework.security.core.context.SecurityContextHolder).context.authentication.name")
   public List<Collection> findAll() {
-    return collectionRepository.findAll(Sort.by(Sort.Direction.ASC, "id"));
+    User currentUser = userService.getCurrentUser();
+    return collectionRepository.findByOwner(currentUser, Sort.by(Sort.Direction.ASC, "id"));
   }
 
-  @Cacheable(value = "collections", key = "#id")
+  @Cacheable(
+      value = "collections",
+      key =
+          "'user:' + T(org.springframework.security.core.context.SecurityContextHolder).context.authentication.name + ':id:' + #id")
   public Optional<Collection> findById(Long id) {
-    return collectionRepository.findById(id);
+    User currentUser = userService.getCurrentUser();
+    return collectionRepository.findByIdAndOwner(id, currentUser);
   }
 
   @CacheEvict(value = "collections", allEntries = true)
   public Collection save(Collection collection) {
+    User currentUser = userService.getCurrentUser();
+    collection.setOwner(currentUser);
     return collectionRepository.save(collection);
   }
 
   @CacheEvict(value = "collections", allEntries = true)
   public void deleteById(Long id) {
-    collectionRepository.deleteById(id);
+    User currentUser = userService.getCurrentUser();
+    Optional<Collection> collection = collectionRepository.findByIdAndOwner(id, currentUser);
+    if (collection.isPresent()) {
+      collectionRepository.deleteById(id);
+    } else {
+      throw new RuntimeException("Collection not found or you don't have permission to delete it");
+    }
   }
 
   public boolean existsByName(String name) {
-    return collectionRepository.existsByNameIgnoreCase(name);
+    User currentUser = userService.getCurrentUser();
+    return collectionRepository.existsByNameIgnoreCaseAndOwner(name, currentUser);
   }
 
-  @Cacheable(value = "collections", key = "'name:' + #name")
+  @Cacheable(
+      value = "collections",
+      key =
+          "'user:' + T(org.springframework.security.core.context.SecurityContextHolder).context.authentication.name + ':name:' + #name")
   public Optional<Collection> findByName(String name) {
-    return collectionRepository.findByNameIgnoreCase(name);
+    User currentUser = userService.getCurrentUser();
+    return collectionRepository.findByNameIgnoreCaseAndOwner(name, currentUser);
   }
 
   @CacheEvict(value = "collections", allEntries = true)
   public Optional<Collection> update(Long id, Collection updatedCollectionData) {
+    User currentUser = userService.getCurrentUser();
     return collectionRepository
-        .findById(id)
+        .findByIdAndOwner(id, currentUser)
         .map(
             existingCollection -> {
               existingCollection.setName(updatedCollectionData.getName());
